requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.18.1
opt-einsum==3.1.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
sentencepiece==0.1.85
six==1.13.0
tensorboard==1.12.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tf-slim==1.0
Werkzeug==0.16.0
wrapt==1.11.2
yapf==0.29.0
